GetQuotes improvements
